requirements.txt
crewai==0.80.0
crewai-tools==0.17.0
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
rich==13.9.4
